fix: cast process.pid() to Int to match AppProcess model

+++ app/src/main/kotlin/com/vulcan/app/runtime/RuntimeEngine.kt
@@ -71,7 +71,7 @@ class RuntimeEngine(private val context: Context) {
 
         return AppProcess(
             appId   = app.id,
-            pid     = process.pid().toLong(),
+            pid     = process.pid().toInt(),
             process = process
         )
     }
